# kardo-swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "KARDO API"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/participations": {
      "post": {
        "tags": [
          "Заявка на участие в отборе: Users."
        ],
        "summary": "Добавление заявки на участие в отборе.",
        "operationId": "createParticipation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewParticipationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Отбор не найден",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Заявка добавлена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipationRequestDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/participations/{participationId}": {
      "get": {
        "tags": [
          "Заявка на участие в отборе: Users."
        ],
        "summary": "Получение заявки на участие в отборе.",
        "operationId": "getParticipationById",
        "parameters": [
          {
            "name": "participationId",
            "in": "path",
            "description": "id заявки",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Заявка не найдена",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Заявка найдена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipationRequestDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Заявка на участие в отборе: Users."
        ],
        "summary": "Удаление заявки на участие в отборе по ИД.",
        "operationId": "deleteParticipationById",
        "parameters": [
          {
            "name": "participationId",
            "in": "path",
            "description": "id заявки",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Заявка не найдена",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Заявка удалена."
          }
        }
      },
      "patch": {
        "tags": [
          "Заявка на участие в отборе: Users."
        ],
        "summary": "Обновление заявки на участие в отборе.",
        "operationId": "updateParticipationById",
        "parameters": [
          {
            "name": "participationId",
            "in": "path",
            "description": "id заявки",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateParticipationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Внутренняя ошибка сервера",
            "content": {

            }
          },
          "404": {
            "description": "Заявка не найдена",
            "content": {

            }
          },
          "401": {
            "description": "Пользователь не авторизован",
            "content": {

            }
          },
          "400": {
            "description": "Запрос составлен некорректно",
            "content": {

            }
          },
          "200": {
            "description": "Заявка обновлена.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipationRequestDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UserShortDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "birthday": {
            "type": "string",
            "format": "date"
          },
          "country": {
            "type": "string"
          },
          "avatarPhoto": {
            "type": "string"
          }
        },
        "description": "Сущность пользователя короткая."
      },
...
      "GrandCompetitionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "hashtag": {
            "type": "string"
          },
          "competitionStart": {
            "type": "string",
            "format": "date"
          },
          "competitionEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "FINISHED",
              "UPCOMING",
              "IN_PROGRESS"
            ]
          },
          "location": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BREAKING",
                "PARKOUR",
                "WORKOUT",
                "KICKSCOOTERING",
                "TRICKING",
                "SKATEBOARDING",
                "BMX",
                "HIP_HOP",
                "DJ_ING",
                "GRAFFITI",
                "WAKEBOARDING",
                "ALL_FIELD"
              ]
            }
          },
          "description": {
            "type": "string"
          }
        },
        "description": "Сущность Гранд-соревнования."
      },
      "OfflineSelectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "competitionDto": {
            "$ref": "#/components/schemas/GrandCompetitionDto"
          },
          "selectionStart": {
            "type": "string",
            "format": "date"
          },
          "selectionEnd": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string",
            "enum": [
              "FINISHED",
              "UPCOMING",
              "IN_PROGRESS"
            ]
          },
          "selectionType": {
            "type": "string",
            "enum": [
              "REGIONAL",
              "NATIONAL"
            ]
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BREAKING",
                "PARKOUR",
                "WORKOUT",
                "KICKSCOOTERING",
                "TRICKING",
                "SKATEBOARDING",
                "BMX",
                "HIP_HOP",
                "DJ_ING",
                "GRAFFITI",
                "WAKEBOARDING",
                "ALL_FIELD"
              ]
            }
          },
          "location": {
            "type": "string"
          }
        },
        "description": "Сущность оффлайн-отбора."
      },
...
      "NewParticipationRequest": {
        "required": [
          "fields",
          "requesterId",
          "selectionId"
        ],
        "type": "object",
        "properties": {
          "selectionId": {
            "type": "integer",
            "format": "int64"
          },
          "requesterId": {
            "type": "integer",
            "format": "int64"
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BREAKING",
                "PARKOUR",
                "WORKOUT",
                "KICKSCOOTERING",
                "TRICKING",
                "SKATEBOARDING",
                "BMX",
                "HIP_HOP",
                "DJ_ING",
                "GRAFFITI",
                "WAKEBOARDING",
                "ALL_FIELD"
              ]
            }
          }
        },
        "description": "Данные добавляемой заявки"
      },
      "ParticipationRequestDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "selectionDto": {
            "$ref": "#/components/schemas/OfflineSelectionDto"
          },
          "requesterDto": {
            "$ref": "#/components/schemas/UserShortDto"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "CONFIRMED",
              "CANCELED"
            ]
          },
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BREAKING",
                "PARKOUR",
                "WORKOUT",
                "KICKSCOOTERING",
                "TRICKING",
                "SKATEBOARDING",
                "BMX",
                "HIP_HOP",
                "DJ_ING",
                "GRAFFITI",
                "WAKEBOARDING",
                "ALL_FIELD"
              ]
            }
          }
        },
        "description": "Сущность заявки на участие в отборе."
      },
...
      "UpdateParticipationRequest": {
        "required": [
          "fields"
        ],
        "type": "object",
        "properties": {
          "fields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "BREAKING",
                "PARKOUR",
                "WORKOUT",
                "KICKSCOOTERING",
                "TRICKING",
                "SKATEBOARDING",
                "BMX",
                "HIP_HOP",
                "DJ_ING",
                "GRAFFITI",
                "WAKEBOARDING",
                "ALL_FIELD"
              ]
            }
          }
        },
        "description": "Данные для обновления заявки"
      },
...
    }
  }
}